requirements.txt
gradio==3.0
numpy==1.24.1
Pillow==9.3.0
opencv-python-headless==4.6.0.66
requests==2.28.2
